Add support for include-what-you-use
`include-what-you-use` or `iwyu` is a clang-based tool that looks at
the code's AST to figure out which symbols need to be included and
with the help of user-defined mappings it suggests the include
files that are actually needed.

This is very nice for the per-operator headers build because it give
you a list of exactly the `ATen/ops` headers needed by the file. You
still need to manually write the include-guards etc. but at least this
automates the most tedious part.

This is sill a first draft and the header mappings aren't perfect
because it will still suggest you include basic c10 components
everywhere instead of taking it transitively from `TensorBase.h`.
However, the does provide some useful mappings and removes bad include
paths from the build system that lead to bad suggestions.

ghstack-source-id: 48d5b61a8fa91f734b39bc3b5f901b98f35576b6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71114