Generate aten_interned_strings.h automatically
This generates aten_interned_strings.h from `native_functions.yaml`
which is more like how it was originally done. The items deleted from
`interned_strings.h` are duplicates that need to be removed in order
for the code to compile, some of the remaining items may still be out
of date but it is fairly benign even if that's the case.

ghstack-source-id: 34d54bd9eab322d51842e9242954307cd8f982ba
Pull Request resolved: https://github.com/pytorch/pytorch/pull/69407