dbr quant: refactor first_call hooks to be standalone
Summary:

In a future PR we need to start recording the DAG
of non-quantizeable ops, in order to properly reason about
function fusion.

This PR refactors the prepare hooks to have the first_call
versions be separate functions.  This ensures that the logic
to record the DAG (the first_call functions) is separate
from the logic used at inference.

Test plan:

```
python test/test_quantization.py -k DBR
```

ghstack-source-id: c012ee27497e0abeb8fc06e3c3d2b513f042b03e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/71324